Fixes #114

+++ scripts/elastic_search/files/s3_file.py
@@ -11,9 +11,13 @@ class S3File:
 	def download(self):
 		print "Downloading data from s3 (https://s3.amazonaws.com/" + self.bucket + "/" + self.filename + " -> " + self.savepath + "/" + self.filename + ") ..."
 		if not os.path.exists(self.savepath):
+			print "Making temp file storage: " + self.savepath
 			os.makedirs(self.savepath)
 		url = "https://s3.amazonaws.com/" + self.bucket + "/" + self.filename
-		urllib.urlretrieve(url, self.savepath + "/" + self.filename)
+		if not os.path.exists(self.savepath + "/" + self.filename):
+			urllib.urlretrieve(url, self.savepath + "/" + self.filename)
+		else:
+			print "File: " + self.savepath + "/" + self.filename + " already exists not downloading"
 		return self.savepath + "/" + self.filename
 
 	def list_files(self):
